ISOKANN.jl

Documentation for ISOKANN.jl

ISOKANN.IsoRun — Type

struct IsoRun{T}

The IsoRun struct represents a configuration for running the Isomolly algorithm.

Fields

nd::Integer: Number of outer data subsampling steps.

nx::Integer: Size of subdata set.

np::Integer: Number of power iterations with the same subdata.

nl::Integer: Number of weight updates with the same power iteration step.

nres::Integer: Resample new data every n outer steps.

ny::Integer: Number of new points to sample.

nk::Integer: Number of Koopman points to sample.

nxmax::Integer: Maximal number of x data points.

sim: Simulation object.

model: Model object.

opt: Optimization algorithm.

data::T: Data object.

losses: Vector to store loss values.

loggers::Vector: Vector of loggers.

ISOKANN.run! — Function

run the given IsoRun object

ISOKANN.MollyLangevin — Type

struct MollyLangevin{S<:Molly.System} <: IsoSimulation

The MollyLangevin struct represents a Langevin dynamics simulation for the Molly package. It contains the system as well as the integration parameters.

Fields

sys::S: The system to be simulated.

temp::Float64: The temperature of the simulation in Kelvin. Default is 298.0 K.

gamma::Float64: The friction coefficient for the Langevin dynamics. Default is 1.0.

dt::Float64: The time step size in picoseconds. Default is 2e-3 ps.

T::Float64: The total simulation time in picoseconds. Default is 2e-1 ps.

n_threads::Int: The number of threads for force computations. Default is 1.

ISOKANN.propagate — Function

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

ISOKANN.PDB_1UAO — Method

Create a Molly system for the small Chignolin protein

ISOKANN.PDB_ACEMD — Method

Create a Molly system for the alanine dipeptide without solvent

ISOKANN.PDB_diala_water — Method

Create a Molly system for the alanine dipeptide with water

ISOKANN.pairnet — Function

Fully connected neural network with layers layers from n to nout dimensions. features allows to pass a featurizer as preprocessor, activation determines the activation function for each but the last layer lastactivation can be used to modify the last layers activation function

ISOKANN.pairnet — Function

Flux neural network model with layers fully connected layers using the corresponding simulations features as first layers

ISOKANN.propagate — Method

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

ISOKANN.reactionpath — Method

compute the reactionpath for the simulation sim starting in x0 along the gradient of the function provided by chi

Optional arguments: extrapolate: walk beyond the interval limits orth: factor for the force orthogonal to the chi gradient solver: the ODE solver to use dt: the timestep size kwargs...: Keyword arguments passed to the solve method

ISOKANN.run! — Method

run the given IsoRun object

ISOKANN.solve — Method

sample a single trajectory for the given system