ISOKANN.jl

Documentation for ISOKANN.jl

Main entry points

Missing docstring.

Missing docstring for IsoRun. Check Documenter's build log for details.

ISOKANN.run! — Function

run the given IsoRun object

ISOKANN.MollyLangevin — Type

struct MollyLangevin{S<:Molly.System} <: IsoSimulation

The MollyLangevin struct represents a Langevin dynamics simulation for the Molly package. It contains the system as well as the integration parameters.

Fields

sys::S: The system to be simulated.

temp::Float64: The temperature of the simulation in Kelvin. Default is 298.0 K.

gamma::Float64: The friction coefficient for the Langevin dynamics. Default is 1.0.

dt::Float64: The time step size in picoseconds. Default is 2e-3 ps.

T::Float64: The total simulation time in picoseconds. Default is 2e-1 ps.

n_threads::Int: The number of threads for force computations. Default is 1.

ISOKANN.propagate — Function

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

Public API

ISOKANN.PDB_1UAO — Method

Create a Molly system for the small Chignolin protein

ISOKANN.PDB_ACEMD — Method

Create a Molly system for the alanine dipeptide without solvent

ISOKANN.PDB_diala_water — Method

Create a Molly system for the alanine dipeptide with water

ISOKANN.pairnet — Function

Fully connected neural network with layers layers from n to nout dimensions. features allows to pass a featurizer as preprocessor, activation determines the activation function for each but the last layer lastactivation can be used to modify the last layers activation function

ISOKANN.pairnet — Function

Flux neural network model with layers fully connected layers using the corresponding simulations features as first layers

ISOKANN.propagate — Method

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

ISOKANN.reactionpath — Method

compute the reactionpath for the simulation sim starting in x0 along the gradient of the function provided by chi

Optional arguments: extrapolate: walk beyond the interval limits orth: factor for the force orthogonal to the chi gradient solver: the ODE solver to use dt: the timestep size kwargs...: Keyword arguments passed to the solve method

ISOKANN.run! — Method

run the given IsoRun object

ISOKANN.solve — Method

sample a single trajectory for the given system

Internal API

ISOKANN._pickclosestloop — Method

scales with n=length(hs)

ISOKANN.adddata — Method

adddata(data::D, model, sim, ny, lastn=1_000_000)::D

Generate new data for ISOKANN by adaptive subsampling using the chi-stratified/-uniform method.

Adaptively subsample ny points from data uniformly along their model values.

propagate according to the simulation model.

return the newly obtained data concatenated to the input data

The subsamples are taken only from the lastn last datapoints in data. If renormalize is true, rescale the model values to the interval [0,1] before subsampling.

Examples

julia> (xs, ys) = adddata((xs,ys), chi, mollysim)

ISOKANN.bootstrap — Method

compute initial data by propagating the molecules initial state to obtain the xs and propagating them further for the ys

ISOKANN.centercoords — Method

centercoords any given states by shifting their individual 3d mean to the origin

ISOKANN.chi_exit_rate — Method

compute the chi exit rate as per Ernst, Weber (2017), chap. 3.3

ISOKANN.diala2 — Method

ISOKANN 2 dialanine experiment generates data from 1d ISOKANN and trains a 3d ISOKANN on it The kwargs are passed on to isotarget

ISOKANN.eigenvalue — Method

estimate the eigenvalue of the chi function at the extrema of the data

ISOKANN.eigenvalue_regression — Method

estimate the eigenvalue of the chi function by regression over all data

ISOKANN.extractdata — Function

save data into a pdb file sorted by model evaluation

ISOKANN.featurizer — Method

returns a function features(x) which map the system coordinates to its features for the Flux model

ISOKANN.fixperm — Method

fixperm(new, old)

Permutes the rows of new such as to minimize L1 distance to old.

Arguments

new: The data to match to the reference data.

old: The reference data.

ISOKANN.flatpairdists — Method

Threaded computation of pairwise dists for x::AbstractArray. Assumes each col of x to be a flattened representation of multiple 3d coords. Returns the flattened pairwise distances as columns.

ISOKANN.getcoords — Method

extract the unitful SVector coords from sys and return as a normal vector

ISOKANN.gpu — Method

move the ::System to the GPU, mirroring behavior of Flux.gpu

ISOKANN.growmodel — Method

Given a model and return a copy with its last layer replaced with given output dimension n

ISOKANN.inputdim — Method

obtain the input dimension of a Flux model

ISOKANN.iso2 — Method

iso2(; n=1000, nx=100, ny=10, nd=2, sys=Doublewell(), lr=1e-2, decay=1e-5)

ISOKANN 2.0 under construction.

Arguments

n: Number of iterations (default: 1000)

nx: Number of start points (default: 100)

ny: Number of end points per start (default: 10)

nd: Number of dimensions of the koopman function (default: 2)

sys: System object (needs to support randx0, dim, propagate) (default: Doublewell())

lr: Learning rate

decay: Decay rate

ISOKANN.isokann — Method

ISOKANN with advanced data handling, dataset handles possible resampling of data

ISOKANN.isokann — Method

wrap the old ISORun into the simplified isokann() call

ISOKANN.isokann_simple — Method

bare bones ISOKANN, for educational purposes model is a Flux neural network, opt an Optimiser, data a tuple of (x,y) coordinates, n the number of iterations

ISOKANN.isosteps — Function

isostep(model, opt, (xs, ys), nkoop=1, nupdate=1)

train the model on a given batch of trajectory data (xs, ys) with

nkoop: outer iterations, i.e. reevaluating Koopman

nupdate: inner iterations, i.e. updating the neural network on fixed data

ISOKANN.isotarget — Method

isotarget(model, xs, ys, method=:inv, kwargs...) compute the isokann target function for the given model and data. optional arguments:

method: :inv or :isa for the pseudoinverse or inner simplex algorithm

normalize: normalize the target function to have unit norm (:inv only)

permute: permute the target function to match the chi function

direct: use the direct method instead of the inverse method (:inv only)

eigenvecs: use the eigenvecs of the schur decomposition instead of the identity matrix (:inv only)

ISOKANN.koopman — Method

evluation of koopman by shiftscale(mean(model(data))) on the data

ISOKANN.learnbatch! — Method

batched supervised learning for a given batchsize

ISOKANN.learnstep! — Method

single supervised learning step

ISOKANN.mean_speed — Method

what is the mean speed for a given force

ISOKANN.outputdim — Method

Obtain the output dimension of a Flux model

ISOKANN.pickclosest_sort — Method

pickclosest(haystack, needles)

Return the indices into haystack which lie closest to needles without duplicates by removing haystack candidates after a match. Note that this is not invariant under pertubations of needles

scales with n log(n) m where n=length(haystack), m=length(needles)

ISOKANN.plotatoms! — Function

scatter plot of all first "O" atoms of the starting points xs as well as the "O" atoms from the koopman samples to the first point from ys

ISOKANN.plotlossdata — Function

combined plot of loss graph and current atoms

ISOKANN.rotationhandles — Method

Return the rotation handle indices for the standardform of a molecule

ISOKANN.scatter_chifix — Method

fixed point plot, i.e. x vs model(x)

ISOKANN.selectrows — Method

given an array of arbitrary shape, select the rows inds in the first dimension

ISOKANN.shiftscale — Method

empirical shift-scale operation

ISOKANN.sigma — Function

return the diagonal of the noise factor sigma of the brownian dynamics

ISOKANN.stratified_x0 — Method

stratified_x0(model, ys, n; renormalize=false)

Given an array of states ys, return n model-stratified subsamples.

If the model returns multiple chi-value, subsample along each dimension. If renormalize is true, rescale the model values to the interval [0,1] before subsampling.

ISOKANN.subsample — Method

subsample n points of data uniformly according to the provided model

Works for points provided either as plain Array or as (xs, ys) data tuple

subsample(model, data::Array, n) :: Matrix subsample(model, data::Tuple, n) :: Tuple

ISOKANN.subsample_uniformgrid — Method

subsbample_uniformgrid(ys, n) -> is

given a list of values ys, return nindicesissuch thatys[is]` are approximately uniform by picking the closest points to a randomly perturbed grid in [0,1].

ISOKANN.vec_to_coords — Method

convert normal vector of coords to SVectors of unitful system coords