ISOKANN.jl

Documentation for ISOKANN.jl

Main entry points

ISOKANN.IsoRun — Type

struct IsoRun{T}

The IsoRun struct represents a configuration for running the ISOKANN algorithm with adaptive sampling.

The whole algorithm consists of three nested loop

nd iterations of the data loop where nx points are subsampled (via stratified χ-subsampling) from the pool of all available data

np iterations of the power iteration where the training target is determined with the current model and subdata

nl iterations of the SGD updates to the neural network model to learn the current target

On initialization it samples ny starting positions with nk Koopman samples each. Furthermore if nres > 0 it samples ny new data points adaptively starting from χ-sampled positions every nres steps in the data loop.

The sim field takes any simulation object that implements the data sampling interface (mainly the propagate method, see data.jl), usually a MollyLangevin simulation.

model and opt store the neural network model and the optimizert (defaulting to a pairnet and AdamRegularized).

data contains the training data and is by default constructed using the bootstrap method.

The vector losses keeps track of the training loss and loggers allows to pass in logging functions which are executed in the power iteration loop.

To start the actual training call the run! method.

Fields

nd::Int64: Number of outer data subsampling steps.

nx::Int64: Size of subdata set.

np::Int64: Number of power iterations with the same subdata.

nl::Int64: Number of weight updates with the same power iteration step.

nres::Int64: Resample new data every n outer steps.

ny::Int64: Number of new points to sample.

nk::Int64: Number of Koopman points to sample.

nxmax::Int64: Maximal number of x data points.

sim: Simulation object.

model: Model object.

opt: Optimization algorithm.

data::T: Data object.

losses: Vector to store loss values.

loggers::Vector: Vector of loggers.

ISOKANN.run! — Function

run the given IsoRun object

ISOKANN.MollyLangevin — Type

struct MollyLangevin{S<:Molly.System} <: IsoSimulation

The MollyLangevin struct represents a Langevin dynamics simulation for the Molly package. It contains the system as well as the integration parameters.

Fields

sys::S: The system to be simulated.

temp::Float64: The temperature of the simulation in Kelvin. Default is 298.0 K.

gamma::Float64: The friction coefficient for the Langevin dynamics. Default is 1.0.

dt::Float64: The time step size in picoseconds. Default is 2e-3 ps.

T::Float64: The total simulation time in picoseconds. Default is 2e-1 ps.

n_threads::Int: The number of threads for force computations. Default is 1.

ISOKANN.propagate — Function

propagate(s::OpenMMSimulation, x0::AbstractMatrix{T}, ny; nthreads=Threads.nthreads(), mmthreads=1) where {T}

Propagates ny replicas of the OpenMMSimulation s from the inintial states x0.

Arguments

s: An instance of the OpenMMSimulation type.

x0: Matrix containing the initial states as columns

ny: The number of replicas to create.

Optional Arguments

nthreads: The number of threads to use for parallelization of multiple simulations.

mmthreads: The number of threads to use for each OpenMM simulation. Set to "gpu" to use the GPU platform.

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

propagate is the main interface facilitating sampling of a system. TODO: specify the actual interface required for a simulation to be runnable by ISOKANN.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

Fallback to simulate MD dynamics on the CPU

Public API

ISOKANN.PDB_1UAO — Method

Create a Molly system for the small Chignolin protein

ISOKANN.PDB_ACEMD — Method

Create a Molly system for the alanine dipeptide without solvent

ISOKANN.PDB_diala_water — Method

Create a Molly system for the alanine dipeptide with water

ISOKANN.iso2 — Method

iso2(; n=1000, nx=100, ny=10, nd=2, sys=Doublewell(), lr=1e-2, decay=1e-5)

ISOKANN 2.0 under construction.

Arguments

n: Number of iterations (default: 1000)

nx: Number of start points (default: 100)

ny: Number of end points per start (default: 10)

nd: Number of dimensions of the koopman function (default: 2)

sys: System object (needs to support randx0, dim, propagate) (default: Doublewell())

lr: Learning rate

decay: Decay rate

ISOKANN.pairnet — Function

Fully connected neural network with layers layers from n to nout dimensions. features allows to pass a featurizer as preprocessor, activation determines the activation function for each but the last layer lastactivation can be used to modify the last layers activation function

ISOKANN.pairnet — Method

A pairnet() model with using the simulation's featureinds() particles pairwise positions as input for the first layer

ISOKANN.propagate — Method

propagate(ms::MollyLangevin, x0::AbstractMatrix, ny)

Burst simulation of the MollyLangeving system ms. Propagates ny samples for each initial position provided in the columns of x0.

propagate is the main interface facilitating sampling of a system. TODO: specify the actual interface required for a simulation to be runnable by ISOKANN.

Arguments

ms::MollyLangevin: The MollyLangevin solver object.

x0::AbstractMatrix: The initial positions matrix.

ny: The number of trajectories per initial condition.

Returns

ys: A 3-dimensional array of size (dim(ms), nx, ny) containing the propagated solutions.

ISOKANN.propagate — Method

Fallback to simulate MD dynamics on the CPU

ISOKANN.reactionpath — Method

compute the reactionpath for the simulation sim starting in x0 along the gradient of the function provided by chi

Optional arguments: extrapolate: walk beyond the interval limits orth: factor for the force orthogonal to the chi gradient solver: the ODE solver to use dt: the timestep size kwargs...: Keyword arguments passed to the solve method

ISOKANN.run! — Method

run the given IsoRun object

ISOKANN.solve — Method

sample a single trajectory for the given system

Internal API

ISOKANN.ChiStratData — Type

dataset type handling chi-stratified subsampling and traindata-subsampling

ISOKANN.DataTuple — Type

DataTuple = Tuple{Matrix{T},Array{T,3}} where {T<:Number}

We represent data as a tuple of xs and ys.

xs is a matrix of size (d, n) where d is the dimension of the system and n the number of samples. ys is a tensor of size (d, n, k) where k is the number of koopman samples.

ISOKANN.MutTuple — Type

a tuple which allows mutation and creation of new fields

ISOKANN.OverdampedLangevinGirsanov — Type

OverdampedLangevinGirsanov(; <keyword arguments>)

Simulates the overdamped Langevin equation using the Euler-Maruyama method with an auxilliary control w/u with σ = sqrt(2KT/(mγ)) dX = (-∇U(X)/(γm) + σu) dt + σ dW

where u is the control function, such that u(x,t) = σ .* w(x,t). The accumulated Girsanov reweighting is stored in the field g

Arguments

dt::S: the time step of the simulation.

temperature::K: the equilibrium temperature of the simulation.

friction::F: the friction coefficient of the simulation.

remove_CM_motion=1: remove the center of mass motion every this number of steps, set to false or 0 to not remove center of mass motion.

w::Function: the control function, such that u(x,t) = σ .* w(x,t)

ISOKANN.TransformISA — Type

TransformISA(permute)

Compute the target via the inner simplex algorithm (without feasiblization routine). permute specifies whether to apply the stabilizing permutation

ISOKANN.TransformPseudoInv — Type

TransformPseudoInv(normalize, direct, eigenvecs, permute)

Compute the target by approximately inverting the action of K with the Moore-Penrose pseudoinverse.

If direct==true solve chi * pinv(K(chi)), otherwise inv(K(chi) * pinv(chi))). eigenvecs specifies whether to use the eigenvectors of the schur matrix. normalize specifies whether to renormalize the resulting target vectors. permute specifies whether to permute the target for stability.

ISOKANN.TransformShiftscale — Type

TransformShiftscale()

Classical 1D shift-scale (ISOKANN 1)

ISOKANN._pickclosestloop — Method

scales with n=length(hs)

ISOKANN.adddata — Method

adddata(data::D, model, sim, ny, lastn=1_000_000)::D

Generate new data for ISOKANN by adaptive subsampling using the chi-stratified/-uniform method.

Adaptively subsample ny points from data uniformly along their model values.

propagate according to the simulation model.

return the newly obtained data concatenated to the input data

The subsamples are taken only from the lastn last datapoints in data.

Examples

julia> (xs, ys) = adddata((xs,ys), chi, mollysim)

ISOKANN.bootstrap — Method

bootstrap(sim, nx, ny) :: DataTuple

compute initial data by propagating the molecules initial state to obtain the xs and propagating them further for the ys

ISOKANN.centercoords — Method

centercoords any given states by shifting their individual 3d mean to the origin

ISOKANN.chi_exit_rate — Method

compute the chi exit rate as per Ernst, Weber (2017), chap. 3.3

ISOKANN.data_from_trajectory — Method

data_from_trajectory(xs::Matrix; reverse=false) :: DataTuple

Generate the lag-1 data from the trajectory xs. If reverse is true, also take the time-reversed lag-1 data.

ISOKANN.diala2 — Method

ISOKANN 2 dialanine experiment generates data from 1d ISOKANN and trains a 3d ISOKANN on it

ISOKANN.eigenvalue — Method

estimate the eigenvalue of the chi function at the extrema of the data

ISOKANN.eigenvalue_regression — Method

estimate the eigenvalue of the chi function by regression over all data

ISOKANN.exportdata — Function

exportdata(data::AbstractArray, model, sys, path="out/data.pdb")

Export data to a PDB file.

This function takes an AbstractArray data, sorts it according to the model evaluation, removes duplicates, transforms it to standard form and saves it as a PDB file to path.

ISOKANN.fixperm — Method

fixperm(new, old)

Permutes the rows of new such as to minimize L1 distance to old.

Arguments

new: The data to match to the reference data.

old: The reference data.

ISOKANN.flatpairdists — Method

Threaded computation of pairwise dists for x::AbstractArray. Assumes each col of x to be a flattened representation of multiple 3d coords. Returns the flattened pairwise distances as columns.

ISOKANN.getcoords — Method

extract the unitful SVector coords from sys and return as a normal vector

ISOKANN.gpu! — Method

gpu!(iso::IsoRun)

move the model and data of the given IsoRun to the GPU for CUDA support

ISOKANN.gpu — Method

move the ::System to the GPU, mirroring behavior of Flux.gpu

ISOKANN.growmodel — Method

Given a model and return a copy with its last layer replaced with given output dimension n

ISOKANN.inputdim — Method

obtain the input dimension of a Flux model

ISOKANN.isokann — Method

ISOKANN with advanced data handling, dataset handles possible resampling of data

ISOKANN.isokann — Method

wrap the old ISORun into the simplified isokann() call

ISOKANN.isokann_simple — Method

bare bones ISOKANN, for educational purposes model is a Flux neural network, opt an Optimiser, data a tuple of (x,y) coordinates, n the number of iterations

ISOKANN.isosteps — Function

isostep(model, opt, (xs, ys), nkoop=1, nupdate=1; transform = TransformISA())

train the model on a given batch of trajectory data (xs, ys) with

nkoop: outer iterations, i.e. reevaluating Koopman

nupdate: inner iterations, i.e. updating the neural network on fixed data

ISOKANN.koopman — Method

evluation of koopman by shiftscale(mean(model(data))) on the data

ISOKANN.learnbatch! — Method

DEPRECATED - batched supervised learning for a given batchsize

ISOKANN.learnstep! — Method

single supervised learning step

ISOKANN.mean_speed — Method

what is the mean speed for a given force

ISOKANN.model_with_opt — Function

convenience wrapper returning the provided model with the default AdamW optimiser

ISOKANN.outputdim — Method

Obtain the output dimension of a Flux model

ISOKANN.pickclosest_sort — Method

pickclosest(haystack, needles)

Return the indices into haystack which lie closest to needles without duplicates by removing haystack candidates after a match. Note that this is not invariant under pertubations of needles

scales with n log(n) m where n=length(haystack), m=length(needles)

ISOKANN.plotatoms! — Function

scatter plot of all first "O" atoms of the starting points xs as well as the "O" atoms from the koopman samples to the first point from ys

ISOKANN.plotlossdata — Function

combined plot of loss graph and current atoms

ISOKANN.rotationhandles — Method

Return the rotation handle indices for the standardform of a molecule

ISOKANN.scatter_chifix — Method

fixed point plot, i.e. x vs model(x)

ISOKANN.selectrows — Method

given an array of arbitrary shape, select the rows inds in the first dimension

ISOKANN.shiftscale — Method

empirical shift-scale operation

ISOKANN.sigma — Function

return the diagonal of the noise factor sigma of the brownian dynamics

ISOKANN.subsample — Method

subsample(model, data::Array, n) :: Matrix
subsample(model, data::Tuple, n) :: Tuple

Subsample n points of data uniformly in model. If model returns multiple values per sample, subsample along each dimension.

ISOKANN.subsample — Method

subsample(data, nx)

return a random subsample of nx points from data

ISOKANN.subsample_inds — Method

subsample_inds(model, xs, n) :: Vector{Int}

Returns n indices of xs such that model(xs[inds]) is approximately uniformly distributed.

ISOKANN.subsample_uniformgrid — Method

subsbample_uniformgrid(ys, n) -> is

given a list of values ys, return nindicesissuch thatys[is]` are approximately uniform by picking the closest points to a randomly perturbed grid in [0,1].

ISOKANN.trajectory — Method

trajectory(sim, nx) generate a trajectory of length nx from the simulation sim

ISOKANN.vec_to_coords — Method

convert normal vector of coords to SVectors of unitful system coords