ISOKANN.jl

Documentation for ISOKANN.jl

Main entry points

Missing docstring.

Missing docstring for IsoRun. Check Documenter's build log for details.

ISOKANN.run! — Function

run!(iso::Iso2, n=1, epochs=1)

Run the training process for the Iso2 model.

Arguments

iso::Iso2: The Iso2 model to train.

n::Int: The number of (outer) Koopman iterations.

epochs::Int: The number of (inner) epochs to train the model for each Koopman evaluation.

Missing docstring.

Missing docstring for MollyLangevin. Check Documenter's build log for details.

ISOKANN.propagate — Function

propagate(s::OpenMMSimulation, x0::AbstractMatrix{T}, ny; nthreads=Threads.nthreads(), mmthreads=1) where {T}

Propagates ny replicas of the OpenMMSimulation s from the inintial states x0.

Arguments

s: An instance of the OpenMMSimulation type.

x0: Matrix containing the initial states as columns

ny: The number of replicas to create.

Optional Arguments

nthreads: The number of threads to use for parallelization of multiple simulations.

mmthreads: The number of threads to use for each OpenMM simulation. Set to "gpu" to use the GPU platform.

Note: For CPU we observed better performance with nthreads = num cpus, mmthreads = 1 then the other way around. With GPU nthreads > 1 should be supported, but on our machine lead to slower performance then nthreads=1.

Public API

ISOKANN.Iso2 — Method

Iso2(data; opt=AdamRegularized(), model=pairnet(data), gpu=false, kwargs...)

ISOKANN.Iso2 — Method

Iso2(sim::IsoSimulation; nx=100, nk=10, nd=1, kwargs...)

Convenience constructor which generates the SimulationData from the simulation sim and constructs the Iso2 object. See also Iso2(data; kwargs...)

Arguments

sim::IsoSimulation: The IsoSimulation object.

nx::Int: The number of starting points.

nk::Int: The number of koopman samples.

nd::Int: Dimension of the χ function.

ISOKANN.SimulationData — Type

struct SimulationData{S,D,C,F}

A struct combining a simulation with the simulated coordinates and corresponding ISOKANN trainingsdata

Fields

sim::S: The simulation object.

data::D: The ISOKANN trainings data.

coords::C: The orginal coordinates of the simulations.

featurizer::F: A function mapping coordinates to ISOKANN features.

ISOKANN.SimulationData — Method

SimulationData(sim::IsoSimulation, nx::Int, nk::Int; ...)
SimulationData(sim::IsoSimulation, xs::AbstractMatrix, nk::Int; ...)
SimulationData(sim::IsoSimulation, (xs,ys); ...)

Generates SimulationData from a simulation with either

nx initial points and nk Koopman samples

xs as initial points and nk Koopman sample

xs as inintial points and ys as Koopman samples

ISOKANN.AdamRegularized — Function

Adam with L2 regularization. Note that this is different from AdamW (Adam+WeightDecay) (c.f. Decay vs L2 Reg.)

ISOKANN.pairnet — Function

Fully connected neural network with layers layers from n to nout dimensions. features allows to pass a featurizer as preprocessor, activation determines the activation function for each but the last layer lastactivation can be used to modify the last layers activation function

ISOKANN.reactionpath_minimum — Function

reactionpath_minimum(iso::Iso2, x0; steps=100)

Compute the reaction path by integrating ∇χ with orthogonal energy minimization.

Arguments

iso::Iso2: The isomer for which the reaction path minimum is to be computed.

x0: The starting point for the reaction path computation.

steps=100: The number of steps to take along the reaction path.

ISOKANN.reactionpath_ode — Method

reactionpath_ode(iso, x0; steps=101, extrapolate=0, orth=0.01, solver=OrdinaryDiffEq.Tsit5(), dt=1e-3, kwargs...)

Compute the reaction path by integrating ∇χ as well as orth * F orthogonal to ∇χ where F is the original force field.

Arguments

iso::Iso2: The isomer for which the reaction path minimum is to be computed.

x0: The starting point for the reaction path computation.

steps=100: The number of steps to take along the reaction path.

minimize=false: Whether to minimize the orthogonal to ∇χ before integration.

extrapolate=0: How fast to extrapolate beyond χ 0 and 1.

orth=0.01: The weight of the orthogonal force field.

solver=OrdinaryDiffEq.Tsit5(): The solver to use for the ODE integration.

dt=1e-3: The initial time step for the ODE integration.

ISOKANN.run! — Function

run!(iso::Iso2, n=1, epochs=1)

Run the training process for the Iso2 model.

Arguments

iso::Iso2: The Iso2 model to train.

n::Int: The number of (outer) Koopman iterations.

epochs::Int: The number of (inner) epochs to train the model for each Koopman evaluation.

ISOKANN.runadaptive! — Method

runadaptive!(iso; generations=1, nx=10, iter=100, cutoff=Inf)

Train iso with adaptive sampling. Sample nx new data points followed by iter isokann iterations and repeat this generations times. cutoff specifies the maximal data size, after which new data overwrites the oldest data.

ISOKANN.save_reactive_path — Function

save_reactive_path(iso::Iso2, coords::AbstractMatrix;
    sigma=1,
    out="out/reactive_path.pdb",
    source,
    kwargs...)

Extract and save the reactive path of a given iso.

Computes the maximum likelihood path with parameter sigma along the given data points, aligns it and saves it to the out path.

Arguments

iso::Iso2: The isomer for which the reactive path is computed.

coords::AbstractMatrix: The coordinates corresponding to the samples in iso

sigma=1: The standard deviation used for the reactive path calculation.

out="out/reactive_path.pdb": The output file path for saving the reactive path.

source: The source .pdb file

= kwargs...: additional parameters passed to reactive_path.

Returns

ids: The IDs of the reactive path.

ISOKANN.OpenMM.OpenMMSimulation — Type

A Simulation wrapping the Python OpenMM Simulation object

ISOKANN.OpenMM.OpenMMSimulation — Method

OpenMMSimulation(; pdb=DEFAULT_PDB, ligand="", forcefields=["amber14-all.xml", "amber14/tip3pfb.xml"], temp=298, friction=1, step=0.002, steps=100, features=nothing, minimize=false)

Constructs an OpenMM simulation object.

Arguments

pdb::String: Path to the PDB file.

ligand::String: Path to ligand file.

forcefields::Vector{String}: List of force field XML files.

temp::Float64: Temperature in Kelvin.

friction::Float64: Friction coefficient in 1/picosecond.

step::Float64: Integration step size in picoseconds.

steps::Int: Number of simulation steps.

features: Which features to use for learning the chi function. - A vector of Int denotes the indices of all atoms to compute the pairwise distances from. - A vector of CartesianIndex{2} computes the specific distances between the atom pairs. - A number denotes the radius below which all pairs of atoms will be used (computed only on the starting configuration) - If nothing all pairwise distances are used.

minimize::Bool: Whether to perform energy minimization on first state.

nthreads: The number of threads to use for parallelization of multiple simulations.

mmthreads: The number of threads to use for each OpenMM simulation. Set to "gpu" to use the GPU platform.

Returns

OpenMMSimulation: An OpenMMSimulation object.

Internal API

ISOKANN.DataTuple — Type

DataTuple = Tuple{Matrix{T},Array{T,3}} where {T<:Number}

We represent data as a tuple of xs and ys.

xs is a matrix of size (d, n) where d is the dimension of the system and n the number of samples. ys is a tensor of size (d, k, n) where k is the number of koopman samples.

ISOKANN.IsoSimulation — Type

abstract type IsoSimulation

Abstract type representing an IsoSimulation. Should implement the methods getcoords, propagate, dim

ISOKANN.Stabilize2 — Type

TransformStabilize(transform, last=nothing)

Wraps another transform and permutes its target to match the previous target

Currently we also have the stablilization (wrt to the model though) inside each Transform. TODO: Decide which to keep

ISOKANN.TransformISA — Type

TransformISA(permute)

Compute the target via the inner simplex algorithm (without feasiblization routine). permute specifies whether to apply the stabilizing permutation

ISOKANN.TransformPseudoInv — Type

TransformPseudoInv(normalize, direct, eigenvecs, permute)

Compute the target by approximately inverting the action of K with the Moore-Penrose pseudoinverse.

If direct==true solve chi * pinv(K(chi)), otherwise inv(K(chi) * pinv(chi))). eigenvecs specifies whether to use the eigenvectors of the schur matrix. normalize specifies whether to renormalize the resulting target vectors. permute specifies whether to permute the target for stability.

ISOKANN.TransformShiftscale — Type

TransformShiftscale()

Classical 1D shift-scale (ISOKANN 1)

Base.merge — Method

merge(d1::SimulationData, d2::SimulationData)

Merge the data and features of d1 and d2, keeping the simulation and features of d1. Note that there is no check if simulation features agree.

ISOKANN._pickclosestloop — Method

scales with n=length(hs)

ISOKANN.adddata — Method

adddata(data::D, model, sim, ny, lastn=1_000_000)::D

Generate new data for ISOKANN by adaptive subsampling using the chi-stratified/-uniform method.

Adaptively subsample ny points from data uniformly along their model values.

propagate according to the simulation model.

return the newly obtained data concatenated to the input data

The subsamples are taken only from the lastn last datapoints in data.

Examples

julia> (xs, ys) = adddata((xs,ys), chi, mollysim)

ISOKANN.adddata — Method

adddata(d::SimulationData, model, n)

χ-stratified subsampling. Select n samples amongst the provided ys/koopman points of d such that their χ-value according to model is approximately uniformly distributed and propagate them. Returns a new SimulationData which has the new data appended.

ISOKANN.addextrapolates! — Method

addextrapolates!(iso, n, stepsize=0.01, steps=10)

Sample new data starting points obtained by extrapolating the chi function beyond the current extrema and attach it to the iso objects data.

Samples n points at the lower and upper end each, resulting in 2n new points. stepis the magnitude of chi-value-change per step andsteps`` is the number of steps to take. E.g. 10 steps of stepsize 0.01 result in a change in chi of about 0.1.

The obtained data is filtered such that unstable simulations should be removed, which may result in less then 2n points being added.

ISOKANN.bootstrap — Method

bootstrap(sim, nx, ny) :: DataTuple

compute initial data by propagating the molecules initial state to obtain the xs and propagating them further for the ys

ISOKANN.centercoords — Method

centercoords any given states by shifting their individual 3d mean to the origin

ISOKANN.chi_exit_rate — Method

compute the chi exit rate as per Ernst, Weber (2017), chap. 3.3

ISOKANN.data_from_trajectory — Method

data_from_trajectory(xs::Matrix; reverse=false) :: DataTuple

Generate the lag-1 data from the trajectory xs. If reverse is true, also take the time-reversed lag-1 data.

ISOKANN.exportdata — Function

exportdata(data::AbstractArray, model, sys, path="out/data.pdb")

Export data to a PDB file.

This function takes an AbstractArray data, sorts it according to the model evaluation, removes duplicates, transforms it to standard form and saves it as a PDB file to path.

ISOKANN.extrapolate — Function

extrapolate(iso, n, stepsize=0.1, steps=1, minimize=true)

Take the n most extreme points of the chi-function of the iso object and extrapolate them by stepsize for steps steps beyond their extrema, resulting in 2n new points. If minimize is true, the new points are energy minimized.

ISOKANN.fixperm — Method

fixperm(new, old)

Permutes the rows of new such as to minimize L1 distance to old.

Arguments

new: The data to match to the reference data.

old: The reference data.

ISOKANN.flatpairdists — Method

flatpairdists(x)

Assumes each col of x to be a flattened representation of multiple 3d coords. Returns the flattened pairwise distances as columns.

ISOKANN.flattenfirst — Method

collapse the first and second dimension of the array A into the first dimension

ISOKANN.growmodel — Method

Given a model and return a copy with its last layer replaced with given output dimension n

ISOKANN.inputdim — Method

obtain the input dimension of a Flux model

ISOKANN.kabsch — Method

compute R such that R*p is closest to q

ISOKANN.koopman — Method

evluation of koopman by shiftscale(mean(model(data))) on the data

ISOKANN.load — Method

load(path::String, iso::Iso2)

Load the Iso2 object from a JLD2 file Note that it will be loaded to the CPU, even if it was saved on the GPU. An OpenMMSimulation will be reconstructed anew from the saved pdb file.

ISOKANN.localpdistinds — Method

localpdistinds(coords::AbstractMatrix, radius)

Given coords of shape ( 3n x frames ) return the pairs of indices whose minimal distance along all frames is at least once lower then radius

ISOKANN.model_with_opt — Function

convenience wrapper returning the provided model with the default AdamW optimiser

ISOKANN.outputdim — Method

Obtain the output dimension of a Flux model

ISOKANN.pairdistfeatures — Method

pairdistfeatures(inds::AbstractVector)

Returns a featurizer function which computes the pairwise distances between the particles specified by inds

ISOKANN.pdists — Method

pdists(coords::AbstractArray, inds::Vector{<:Tuple})

Compute the pairwise distances between the particles specified by the tuples inds over all frames in traj. Assumes a column contains all 3n coordinates.

ISOKANN.pickclosest_sort — Method

pickclosest(haystack, needles)

Return the indices into haystack which lie closest to needles without duplicates by removing haystack candidates after a match. Note that this is not invariant under pertubations of needles

scales with n log(n) m where n=length(haystack), m=length(needles)

ISOKANN.plotatoms! — Function

scatter plot of all first "O" atoms of the starting points xs as well as the "O" atoms from the koopman samples to the first point from ys

ISOKANN.plotlossdata — Function

combined plot of loss graph and current atoms

ISOKANN.reactionforce — Function

reactionforce(iso, sim, x, direction, orth=1)

Compute the vector f with colinear component to dchi/dx such that dchi/dx * f = 1 and orth*forcefield in the orthogonal space

ISOKANN.save — Method

save(path::String, iso::Iso2)

Save the complete Iso2 object to a JLD2 file

ISOKANN.savecoords — Function

savecoords(path::String, iso::Iso2, inds=1:numobs(iso.data))

Save the coordinates of the specified observation indices from the data of of iso to the file path.

savecoords(path::String, iso::Iso2, coords::AbstractArray)

Save the coordinates of the specified matrix of coordinates to a file, using the molecule in iso as a template.

ISOKANN.saveextrema — Method

saveextrema(path::String, iso::Iso2)

Save the two extermal configurations (metastabilities) to the file path.

ISOKANN.scatter_chifix — Method

fixed point plot, i.e. x vs model(x)

ISOKANN.selectrows — Method

given an array of arbitrary shape, select the rows inds in the first dimension

ISOKANN.shiftscale — Method

empirical shift-scale operation

ISOKANN.simulationtime — Method

simulationtime(iso::Iso2)

print and return the total simulation time contained in the data of iso in nanoseconds.

ISOKANN.sqpairdist — Method

simplepairdists(x::AbstractArray{<:Any,3})

Compute the pairwise distances between the columns of x, batched along the 3rd dimension.

ISOKANN.subsample — Method

subsample(model, data::Array, n) :: Matrix
subsample(model, data::Tuple, n) :: Tuple

Subsample n points of data uniformly in model. If model returns multiple values per sample, subsample along each dimension.

ISOKANN.subsample — Method

subsample(data, nx)

return a random subsample of nx points from data

ISOKANN.subsample_inds — Method

subsample_inds(model, xs, n) :: Vector{Int}

Returns n indices of xs such that model(xs[inds]) is approximately uniformly distributed.

ISOKANN.subsample_uniformgrid — Method

subsbample_uniformgrid(ys, n) -> is

given a list of values ys, return nindicesissuch thatys[is]` are approximately uniform by picking the closest points to a randomly perturbed grid in [0,1].

ISOKANN.trajectory — Method

trajectory(sim, nx) generate a trajectory of length nx from the simulation sim

ISOKANN.OpenMM.set_random_velocities! — Method

mutates the state in sim

ISOKANN.OpenMM.trajectory — Method

trajectory(s::OpenMMSimulation, x0, steps=s.steps, saveevery=1; stepsize = s.step, mmthreads = s.mmthreads)

Return the coordinates of a single trajectory started at x0 for the given number of steps where each saveevery step is stored.

ISOKANN.propagate — Method

propagate(s::OpenMMSimulation, x0::AbstractMatrix{T}, ny; nthreads=Threads.nthreads(), mmthreads=1) where {T}

Propagates ny replicas of the OpenMMSimulation s from the inintial states x0.

Arguments

s: An instance of the OpenMMSimulation type.

x0: Matrix containing the initial states as columns

ny: The number of replicas to create.

Optional Arguments

nthreads: The number of threads to use for parallelization of multiple simulations.

mmthreads: The number of threads to use for each OpenMM simulation. Set to "gpu" to use the GPU platform.

Note: For CPU we observed better performance with nthreads = num cpus, mmthreads = 1 then the other way around. With GPU nthreads > 1 should be supported, but on our machine lead to slower performance then nthreads=1.

ISOKANN.randx0 — Method

generate n random inintial points for the simulation mm